% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of accidents by state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric. The state number.}

\item{year}{Numeric. The relevant year.}
}
\value{
\code{fars_map_state} returns a plot of the map of the given state, as well
  as the geographic location of the accidents occurring in that state for the given year.
  If the state number does not exist, \code{fars_map_state} will throw an error.
  If there are no accidents in the state for the given year, \code{fars_map_state} will
  display a message indicating this.
}
\description{
\code{fars_map_state} plots the geographic location of accidents for the
input state and year.
}
\examples{
\dontrun{fars_map_state(23, 2014)}

\dontrun{fars_map_state(52, 2014)}

}
