% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read in a Fatality Analysis Reporting System data file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character. The name of the file containing the data.}
}
\value{
A dataframe table.
  If \code{filename} does not exist, an error will be thrown.
}
\description{
\code{fars_read} reads in a data file from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System.
}
\examples{
\dontrun{fars_read("data/accident_2013.csv.bz2")}

}
