% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple Fatality Analysis Reporting System data files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Numeric. A vector of years of the associated files to be read in.}
}
\value{
\code{fars_read_years} will search for the filenames based on the years input.
  Refer to \code{\link{make_filename}} for the file naming conventions used.
  If the files exist a list containing the appropriate data will be returned.
  For files that do not exist, \code{fars_read_years} will throw a warning stating
  the relevant year.
}
\description{
\code{fars_read_years} will read in multiple Fatality Analysis Reporting
System data files based on the years input.
}
\examples{
\dontrun{fars_read_years(2013:2014)}

\dontrun{fars_read_years(2020)}

}
