% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multiple FARS data files by year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Numeric. A vector of years associated with the files to be summarized.}
}
\value{
\code{fars_summarize_years} returns a dataframe.
  If a file does not exist for a given year, \code{fars_summarize_years}
  will throw a warning regarding the given year.
}
\description{
\code{fars_summarize_years} provides a summary of the number of observations
in the FARS data files for each month for the given years input.
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}

}
