% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric. The year to append to the file name.}
}
\value{
\code{make_filename} will return a file name based on the year input.
  Non-integer inputs are truncated.
}
\description{
\code{make_filename} creates a file name for an accident file using
the year input.
}
\examples{
\dontrun{make_filename(2013)}

}
